cli: add options to skip some time at the start/end of VOD streams
In lieu of proper seek support, or those not wishing to use the HTTP
server, this adds support to skip some time from the beginning or end of
a VOD stream. There are two new parameters;
- `--offset-start`: time offset from the start of the stream to start
                    at
- `--offset-end`: time offset from the end of the stream to stop at

Each parameter takes an argument in the form of `00h00m00s`, each of h
(hours), m (minutes), or s (seconds) is optional, but they must appear
in that order. eg.
- `1m` is 1 minute
- `50s` is 50 seconds
- `1m30s` is 1 minute 30 seconds
- `1h` is 1 hour
- `2h30m10s` is 2 hours, 30 minutes and 10 seconds

Currently only HLS and RTMP streams support these offset options.