stream.hls: mux external subtitle streams in HLS streams
This feature can be enabled by using the `--hls-subtitles` command line
option. Some extra options were added for ffmpeg to allows for some
work arounds when using different versions of ffmpeg.

HLS streams use VTT subtitles, which are text based, MPEG-TS only
supports bitmap based subtitles and ffmpeg cannot convert from text to
bitmap. Because of this the Matroska container format must be used (as
it supports text subtitles), however, MPEG-TS and Matroska support
different types of AAC streams. Because of _this_ a bitstream filter
(bsf) needs to be applied to the audio stream to make it compatible
with Matroska.
Older versions of ffmpeg do not support AAC fully and you must use the
`-strict -2` option, which is why the `--ffmpeg-strict` option was
added.

Fixes #1847
Fixes #1940